Livestatus: fix rounding bug in double columns
Filters on double columns truncated the reference
values to integer, so
Filter: execution_time < 0.5
didn't show no data.

+++ livestatus/src/DoubleColumnFilter.cc
@@ -39,7 +39,7 @@ DoubleColumnFilter::DoubleColumnFilter(DoubleColumn *column, int opid, char *val
 bool DoubleColumnFilter::accepts(void *data)
 {
    bool pass = true;
-   int32_t act_value = _column->getValue(data);
+   double act_value = _column->getValue(data);
    switch (_opid) {
       case OP_EQUAL:
 	 pass = act_value == _ref_value; break;
